+++ changelog.md
@@ -1,6 +1,7 @@
 # Change Log of SilverShop JSON Response
 
 * 3.0.0 update for SilverShop v3
-  * Added Modifiers to response data
+  * Added the current Cart to all response data
+  * Variations appear under `subtitle` key (previously `variations`)
 * 2.0.0 Update for SilverShop v2
 * 1.0.0 Initial Release